feat: Server logging polish and cleanup.
