+++ packages/server/lib/client/Page.js
@@ -21,10 +21,10 @@ import Client from '@lowdefy/client';
 import Head from 'next/head';
 import Link from 'next/link';
 
-import actions from '../build/plugins/actions.js';
-import blocks from '../build/plugins/blocks.js';
-import icons from '../build/plugins/icons.js';
-import operators from '../build/plugins/operators/client.js';
+import actions from '../../build/plugins/actions.js';
+import blocks from '../../build/plugins/blocks.js';
+import icons from '../../build/plugins/icons.js';
+import operators from '../../build/plugins/operators/client.js';
 
 const Page = ({ auth, lowdefy, pageConfig, rootConfig }) => {
   const router = useRouter();
