+++ packages/server/pages/[pageId].js
@@ -16,12 +16,13 @@
 
 import { getPageConfig, getRootConfig } from '@lowdefy/api';
 
-import serverSidePropsWrapper from '../lib/serverSidePropsWrapper.js';
-import Page from '../lib/Page.js';
+import serverSidePropsWrapper from '../lib/server/serverSidePropsWrapper.js';
+import Page from '../lib/client/Page.js';
 
 async function getServerSidePropsHandler({ context, nextContext }) {
   const { pageId } = nextContext.params;
   const { logger } = context;
+  // throw new Error('Test', { cause: { a: 4, pageId } });
   const [rootConfig, pageConfig] = await Promise.all([
     getRootConfig(context),
     getPageConfig(context, { pageId }),
