+++ packages/server/pages/api/request/[pageId]/[requestId].js
@@ -16,7 +16,7 @@
 
 import { callRequest } from '@lowdefy/api';
 
-import apiWrapper from '../../../../lib/apiWrapper.js';
+import apiWrapper from '../../../../lib/server/apiWrapper.js';
 
 async function handler({ context, req, res }) {
   if (req.method !== 'POST') {
@@ -25,7 +25,6 @@ async function handler({ context, req, res }) {
   const { pageId, requestId } = req.query;
   const { blockId, payload } = req.body;
   context.logger.info({ event: 'call_request', pageId, requestId, blockId });
-  context.logger.debug({ payload });
   const response = await callRequest(context, { blockId, pageId, payload, requestId });
   res.status(200).json(response);
 }
